[3694][FIX] stock_picking_accounting_date
Before this commit, accounting date set in inventory adjustments would be nullified
as the date of the account move would be overriden by the accounting date field of
the originating stock move, which was missing consideration on the force_period_date
context.

This commit fixes the issue by respecting the context value in assigning the accounting
date in stock moves.